roll back node manager split
